requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
click==8.1.7
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
prometheus-client==0.20.0
pytest==8.2.0
pytest-asyncio==0.23.7
docker==7.1.0
